assignment2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for simple user store webapp on EC2 for Cloud Computing course. "
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.128.1",
    "pwdlib[argon2]>=0.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
